+++ tests/testthat/test-defaultPlotConfig.R
@@ -18,7 +18,7 @@ test_that("getDefaultPlotConfig returns expected defaults", {
   config <- getDefaultPlotConfig()
 
   expect_true(is.list(config))
-  expect_equal(length(config), 105) # Check number of default parameters
+  expect_equal(length(config), 110) # Check number of default parameters
 
   expect_equal(config$apply_default_scales, TRUE)
   expect_equal(config$apply_default_theme, TRUE)