+++ NEWS.md
@@ -13,6 +13,7 @@
 * Added unit tests for `getDefaultPlotConfig` and `ggPhenotypebyDegree`
 * Added workaround for plotly non-support for geom_curve with computeCurvedMidpoint
 * Adding twin plotting features
+* Added ability to fill colors as a function of the degree of relatedness to a focal person
 
 # ggpedigree 0.4.1
 ## Status: Submitted to CRAN